Renamed Login use case package, fixed tests
