+++ src/main/java/screens/MainEmployeeScreen.java
@@ -9,7 +9,6 @@ import interface_adaptors.update_price.UpdateIview;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
-import use_cases.LoginCheckUserExists.LoginOutPutDS;
 
 public class MainEmployeeScreen extends JPanel {
